+++ packages/app/src/entities/board/index.ts
@@ -1,0 +1,2 @@
+export * from "./board.model";
+export * from "./board.logic";
