+++ packages/app/src/entities/cell/index.ts
@@ -1,0 +1,2 @@
+export * from "./cell.model";
+export * from "./cell.logic";
