+++ packages/app/src/entities/index.ts
@@ -1,0 +1,3 @@
+export * from "./board";
+export * from "./cell";
+export * from "./player";
