+++ packages/app/src/entities/player/index.ts
@@ -1,0 +1,2 @@
+export * from "./player.model";
+export * from "./player.logic";
